pyproject.toml
[tool.poetry]
name = "flask-contacts-demo"
version = "0.1.0"
description = "A demo application in Flask for a workshop during the EPHEC hackaton"
authors = ["Nicolas Surleraux <nsurleraux gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Flask = "^2.0.2"
Flask-Cors = "^3.0.10"
pyhumps = "^3.0.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
flake8 = "^4.0.1"
black = {version = "^21.9b0", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
